unaligned and aligned bitpacking kernels
